% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_dosage.R
\name{snp_dosage}
\alias{snp_dosage}
\title{\code{snp_dosage} extracts SNP dosages from a VCF file given RSID.}
\usage{
snp_dosage(rsid, map, vcf_path, n = 1)
}
\arguments{
\item{rsid}{A string containing the SNP rsid}

\item{map}{A data.frame containing ID, CHR, POS (rsid, chromosome, position)}

\item{vcf_path}{Path to vcf file, including prefix, excluding ".dose.vcf.gz"}

\item{n}{maximum sample size (optional, but useful in case of missing SNPs)}
}
\value{
A data.frame containing allele dosages for the SNP
}
\description{
\code{snp_dosage} extracts SNP dosages from a VCF file given RSID.
}
\examples{
# Example how to extract APOE SNP dosages
# from Gen-R HRC 1.1 imputed data
library(indepth)

# Read in HRC map file to lookup coordinates of SNPs
# from http://www.haplotype-reference-consortium.org/site
map <- fread("HRC.r1-1.GRCh37.wgs.mac5.sites.tab.gz")
setnames(map, "#CHROM", "CHR") #change names

# APOE SNP RS IDs
snp <- c("rs429358","rs7412")

# Path to vcf file
vcf_path <- "~/GENR3/Imputed/HRC_release1.1/chr"

# Extract SNP dosage for all SNPs specified in SNP
dosage.list <- lapply(snp, function(x) snp_dosage(x, map, vcf_path, n = 5756))

# Combine list of SNP dosage
dosage.data <- do.call("cbind", dosage.list)

# Remove the GENR3_ part from participant identifier
# to be consistent with other Gen-R data
dosage.data$IDC <- gsub("GENR3_", "", row.names(dosage.data))

}
