% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_info.R
\name{snp_info}
\alias{snp_info}
\title{\code{snp_info} extracts SNP information}
\usage{
snp_info(rsid, map, vcf_path)
}
\arguments{
\item{rsid}{A string containing the SNP rsid}

\item{map}{A data.frame containing ID, CHR, POS (rsid, chromosome, position)}

\item{vcf_path}{Path to vcf file, including prefix, excluding ".dose.vcf.gz"}
}
\value{
A data.frame containing effect/reference allele,
effect allele frequency and imputation quality (R^2)
}
\description{
\code{snp_info} extracts SNP information
}
\examples{
# Example how to extract APOE SNP dosages and info
# from Gen-R HRC 1.1 imputed data

library(indepth)

# Read in HRC map file to lookup coordinates of SNPs
# from http://www.haplotype-reference-consortium.org/site
map <- fread("HRC.r1-1.GRCh37.wgs.mac5.sites.tab.gz")
setnames(map, "#CHROM", "CHR") #change names

# APOE SNP RS IDs
snp <- c("rs429358","rs7412")

# Path to vcf file
vcf_path <- "~/GENR3/Imputed/HRC_release1.1/chr"

# Extract SNP info for all SNPs specified in SNP
info.data <- lapply(snp, function(x) snp_info(x, map, vcf_path))

# Combine list of SNP dosage
info.data <- do.call("rbind", info.data)

}
